Add viewModel, toolbar, search file, pdf library

+++ app/src/main/java/com/olegdev/documentcreator/models/BaseFile.kt
@@ -5,7 +5,7 @@ import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-open class BaseFile(open var id: Long = 0,
+open class BaseFile(open var id: Int = 0,
                     open var name: String,
                     open var path: Uri
 ) : Parcelable
